set up passport jwt and create autherization and authentication using jwt strategy

+++ controllers/api/v1/posts_api.js
@@ -20,12 +20,18 @@ module.exports.destroy = async (req, res) => {
     try{
         let post = await Post.findById(req.params.id)
 
-        post.remove;
+        if(post.user == req.user.id){
+            post.remove();
     
-        await Comment.deleteMany({post: req.params.id});
-        return res.json(200, {
-            message: "Post and associated comments deleted successfully!"
-        })
+            await Comment.deleteMany({post: req.params.id});
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post"
+            });
+        }
     }catch{
         return res.json(500, {
             message: "Internal Server Error"
